feat(tools): added print cmds, fcn reorg
