+++ docs/cli-commands/lula_tools.md
@@ -24,6 +24,7 @@ Collection of additional commands to make OSCAL easier
 * [lula](./lula.md)	 - Risk Management as Code
 * [lula tools compose](./lula_tools_compose.md)	 - compose an OSCAL component definition
 * [lula tools lint](./lula_tools_lint.md)	 - Validate OSCAL against schema
+* [lula tools print](./lula_tools_print.md)	 - Print Resources or Lula Validation from an Assessment Observation
 * [lula tools template](./lula_tools_template.md)	 - Template an artifact
 * [lula tools upgrade](./lula_tools_upgrade.md)	 - Upgrade OSCAL document to a new version if possible.
 * [lula tools uuidgen](./lula_tools_uuidgen.md)	 - Generate a UUID
