+++ src/cmd/common/common.go
@@ -2,9 +2,12 @@ package common
 
 import (
 	"fmt"
+	"os"
 	"strings"
 
 	"github.com/defenseunicorns/lula/src/internal/template"
+	"github.com/defenseunicorns/lula/src/pkg/message"
+	"github.com/defenseunicorns/lula/src/types"
 )
 
 func ParseTemplateOverrides(setFlags []string) (map[string]string, error) {
@@ -23,4 +26,20 @@ func ParseTemplateOverrides(setFlags []string) (map[string]string, error) {
 		overrides[path] = value
 	}
 	return overrides, nil
+}
+
+// writeResources writes the resources to a file or stdout
+func WriteResources(data types.DomainResources, filepath string) error {
+	jsonData := message.JSONValue(data)
+
+	// If a filepath is provided, write the JSON data to the file.
+	if filepath != "" {
+		err := os.WriteFile(filepath, []byte(jsonData), 0600)
+		if err != nil {
+			return fmt.Errorf("error writing resource JSON to file: %v", err)
+		}
+	} else {
+		message.Printf("%s", jsonData)
+	}
+	return nil
 }
