+++ src/cmd/dev/common.go
@@ -4,11 +4,9 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"os"
 	"strings"
 	"time"
 
-	oscalTypes_1_1_2 "github.com/defenseunicorns/go-oscal/src/types/oscal-1-1-2"
 	"github.com/spf13/cobra"
 	"sigs.k8s.io/yaml"
 
@@ -106,38 +104,4 @@ func RunSingleValidation(ctx context.Context, validationBytes []byte, opts ...ty
 	}
 
 	return lulaValidation, nil
-}
-
-// GetObservationByUuid returns the observation with the given UUID
-func GetObservationByUuid(assessmentResults *oscalTypes_1_1_2.AssessmentResults, observationUuid string) (*oscalTypes_1_1_2.Observation, error) {
-	if assessmentResults == nil {
-		return nil, fmt.Errorf("assessment results is nil")
-	}
-
-	for _, result := range assessmentResults.Results {
-		if result.Observations != nil {
-			for _, observation := range *result.Observations {
-				if observation.UUID == observationUuid {
-					return &observation, nil
-				}
-			}
-		}
-	}
-	return nil, fmt.Errorf("observation with uuid %s not found", observationUuid)
-}
-
-// writeResources writes the resources to a file or stdout
-func writeResources(data types.DomainResources, filepath string) error {
-	jsonData := message.JSONValue(data)
-
-	// If a filepath is provided, write the JSON data to the file.
-	if filepath != "" {
-		err := os.WriteFile(filepath, []byte(jsonData), 0600)
-		if err != nil {
-			return fmt.Errorf("error writing resource JSON to file: %v", err)
-		}
-	} else {
-		message.Printf("%s", jsonData)
-	}
-	return nil
 }
