+++ src/test/e2e/cmd/testdata/tools/print/resources.golden
