+++ src/test/e2e/cmd/testdata/tools/print/validation.golden
